Added timestamp field to all stored events.
This field represents the time the event is added. The field is used in
the context of searches. When a user searches for events, the current
time in milliseconds is added to the search request as parameter. The
query will then execute with a filter event.timestamp < given timestamp.
This way we prevent event added after the initial search popping up
during scrolling through events.