Configurable parser source.
Besides the payload attribute metadata.* attributes can now be used as
source for parsers. The source for parsers can be configured at the
endpoints configuration page. Also a new parser 'Copy value' is
introduced. This parsers makes it possible to set the id and
correlationd_id from values in the metadata.* attribute. If, for example,
a user is copying all http headers to metadata.* attributes the 'Copy
value' parser can copy the X-Request-ID metadata attribute to the id.